# Import bundled Insect in `index.js`