@sharkdp 2022-07-20
> This improves the startup time on my machine significantly. I'll let the benchmark results speak for themselves:

Fantastic :partying_face: Thank you so much!

> Also, this commit causes `output` to not be published to NPM, which makes for a much installation size: Insect 5.7.0 on NPM has an unpacked size of 29.2 MBs, but `npm publish --dry-run` with this commit reports 378.8 kB. `.npmignore` was also removed, as the `files` property in `package.json` includes only the files we want, so `.npmignore` does basically nothing. Read https://docs.npmjs.com/cli/v8/configuring-npm/package-json/#files for more information.

Nice!

I need to look this up somewhere (issue tracker, Git log) but I believe I remember that I tried this in the past (remove the `output` folder) and it caused some problems for people that wanted to install Insect without having to recompile everything from (PureScript) source?