@sharkdp 2022-07-20
Maybe retarget the PR towards your spago branch?